pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-cli"
version = "0.1.0"
description = "A CLI tool for managing multiple AI models like Qwen, Claude, Gemini, etc."
readme = "README.md"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer>=0.9.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.21.0",
    "google-generativeai>=0.4.0",
    "dashscope>=1.13.0"
]
requires-python = ">=3.8"

[project.scripts]
ai-cli = "ai_cli.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
]

[tool.setuptools]
packages = ["ai_cli"]